requirements/base.txt
amqp==2.3.2
celery
coreapi
coreschema
Django
django-rest-auth
django-allauth
django-cors-headers==2.4.0
django-debug-toolbar
django-extensions==2.1.4
django-crispy-forms==1.7.2
django-filter==2.0.0
django-redis==4.10.0
django-smalluuid
djangorestframework
djangorestframework-jwt
drf-yasg
Pillow
kombu==4.2.1
oauth
PyJWT
whitenoise
